# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Sky Holidays API",
        "description": "API documentation for Open Sky Holidays",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/about-our-core-values/{aboutOurCoreValue}": {
            "get": {
                "tags": [
                    "About Core Values"
                ],
                "summary": "Get single about core value details",
                "description": "Retrieves single about core value record by ID.",
                "operationId": "67d83fff8e4cac1174ff151ad650ae40",
                "parameters": [
                    {
                        "name": "aboutOurCoreValue",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Core value retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Core value retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AboutOurCoreValue"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Core value not found"
                    }
                }
            }
        },
...
        "/api/about-why-choose-us/active": {
            "get": {
                "tags": [
                    "About Why Choose Us"
                ],
                "summary": "Get active About Why Choose Us section for website",
                "description": "Retrieves the latest active About Why Choose Us section record for the frontend.",
                "operationId": "f1b59a779896fc7d0a0a6e3ed6c26ab1",
                "responses": {
                    "200": {
                        "description": "Active About Why Choose Us retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Active About Why Choose Us retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AboutWhyChooseUs"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Active About Why Choose Us not found"
                    }
                }
            }
        },
        "/api/about-why-choose-us/{aboutWhyChooseUs}": {
            "get": {
                "tags": [
                    "About Why Choose Us"
                ],
                "summary": "Get About Why Choose Us details",
                "description": "Retrieves single About Why Choose Us section details by ID.",
                "operationId": "4a3c56c3ba06de7d30c6357e5132cbf5",
                "parameters": [
                    {
                        "name": "aboutWhyChooseUs",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "About Why Choose Us retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "About Why Choose Us retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AboutWhyChooseUs"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Section not found"
                    }
                }
            }
        },
...
        "/api/adventure-categories/{adventureCategory}": {
            "get": {
                "tags": [
                    "Adventures"
                ],
                "summary": "Get single adventure category",
                "description": "Retrieves single adventure category by ID.",
                "operationId": "59418fb3e4319f35f1eae93ce6161d59",
                "parameters": [
                    {
                        "name": "adventureCategory",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adventure category retrieved successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AboutOurCoreValue": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Integrity"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa-solid fa-handshake",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "We operate with complete honesty — no hidden fees, no misleading itineraries, just clear and ethical service."
                    }
                },
                "type": "object"
            },
            "AboutWhyChooseUs": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Why Choose Us"
                    },
                    "title": {
                        "type": "string",
                        "example": "Setting Standard for Trust and Comfort."
                    },
                    "description": {
                        "type": "string",
                        "example": "We believe that traveling shouldn't be stressful. We ensure every segment of your journey is organized with precise dedication."
                    },
                    "image": {
                        "description": "Why choose us section image (JPG, JPEG, PNG, WEBP, AVIF)",
                        "type": "string",
                        "example": "about_why_choose_us/photo.jpg",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "Full URL to why choose us image",
                        "type": "string",
                        "example": "https://openskyholidays.com/storage/about_why_choose_us/photo.jpg",
                        "nullable": true
                    },
                    "features_icon": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "fa-solid fa-headset"
                        }
                    },
                    "features_title": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "24/7 Expert Support"
                        }
                    },
                    "features_description": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Our travel assistants are always available to help you navigate queries or itinerary shifts."
                        }
                    },
                    "badge_title": {
                        "type": "string",
                        "example": "Trusted by 15,000+"
                    },
                    "badge_subtitle": {
                        "type": "string",
                        "example": "Happy travelers worldwide"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}